+++ lib/carto/renderer_js.js
@@ -213,7 +213,7 @@ CartoCSS.prototype = {
             rule.toXML(parse_env, {});
         } 
 
-        if (parse_env.errors.message) {
+        if (this.options.strict && parse_env.errors.message) {
           throw new Error(parse_env.errors.message);
         }
 
@@ -282,6 +282,7 @@ carto.RendererJS = function (options) {
     var reference = this.options.reference || require('./torque-reference').version.latest;
     tree.Reference.setData(reference);
     this.options.mapnik_version = this.options.mapnik_version || 'latest';
+    this.options.strict = this.options.hasOwnProperty('strict') ? this.options.strict : false;
 };
 
 // Prepare a javascript object which contains the layers
